requirements.txt
crewai
crewai_tools
python-dotenv
pypdf2
streamlit
langchain
langchain_core
langchain_groq